pkg/pillar: Parse the PinCpu parameter coming from a controller.
The PinCpu parameter sent by a controller reflects the fact, that VM must or
must not pin the VCPUs to CPUs and limit the CPU set used by the Application.
Later it's used in Domain Manager to pass the proper arguments to containerd
and QEMU.
Add the corresponding parameter to the VmConfig structure.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>