pkg/pillar: Implement the setCgroupCpuset function.
The function is used to change the CPU set of the cgroup created to run a VM in
it.
The function is writing to the control file of the corresponding cgroup,
mounted in the `/proc/fs` file system. As the function needs access to the
`/proc/fs` file system, that is accessible only from the host file system only,
we use the `chroot /hostfs` trick here.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>