pkg/pillar: Assign CPUs to an Application on its creation.
For the Applications with the CPUsPinned parameter set to `false`, all the CPUs
not pinned to any other Application are assigned.
For the Applications with the CPUsPinned parameter set to `true` the CPUs are
picked from the list of available not-yet pinned CPUs. After the CPUs are
assigned to the Application, they are excluded from the CPU sets of the
Applications with not-pinned CPUs.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>