pkg/pillar: Assign CPUs to an Application on its restart.
The CPUs assigned to an Application should appear in its Configuration before
the corresponding Domain is (re)created.
During the Application creation, EVE fills the CPUs field of the Configuration
and also stores it in the DomainContext structure.
During the Application restart, EVE restores the CPUs field value from the
previously stored values in the DomainContext structure

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>