pkg/pillar: Lock access to the CPUs information stored in DomainContext.
DomainContext can be accessed concurrently by several threads. Some of them may
try to (re)assign the CPUs to the Applications. Not to have conflicts, EVE have
to lock access to the corresponding fields.
The lock is taken when EVE wants to fix the information about the CPUs usage.
The lock is released when the information is fixed and ready to be passed as
arguments to containerd and QEMU.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>