zedmanager: Extract the list of the requested snapshots.
Retrieve the list of snapshots that first appear in the application
configuration and save them in the SnapshotsToBeTaken list of the
AppInstanceStatus struct. This ensures that they can be taken when
necessary.

The changes introduce two new functions: getNewSnapshotRequests and
isNewSnapshotRequest. GetNewSnapshotRequests retrieves the list of new
snapshot requests, while isNewSnapshotRequest checks whether a snapshot
is already present in either SnapshotsToBeTaken or AvailableSnapshots
lists.

Additionally, the code in handleModify has been updated to make use of
the new SnapshotsToBeTaken list to handle snapshot creation during
application updates.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>