pillar/zedmanager: Update command counter checks to handle snapshot rollbacks.
Change the condition for RestartCmd and PurgeCmd from equality to a
strict greater than comparison. This is to account for cases where the
counter might decrease during a snapshot rollback and the application of
the old config. This change ensures that the appropriate command is
triggered only when the new counter is strictly greater than the old
one.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>